webrtc py package name update
I believe the webrtc module name has changes to webrtcvad-wheels. command to run it is `pip3 install webrtcvad-wheels`. I had to run this to make it work. I apologize for not sharing the error message. Please let me know if you are able to test it. It is alright if you are not able to test. It doesn't affect it much anyways, so we can leave it without updating it.